\name{dbConnectGUI}
\alias{dbConnectGUI}
\title{A Gui to connect to a MySQL database}
\usage{
  dbConnectGUI(connection = NULL)
}
\arguments{
  \item{connection}{An option connection created by
  getConnection.}
}
\description{
  A GUI to connect to a MySQL database
}
\examples{
\dontrun{
dbConnectGUI()
}
}

