\name{sqlToR}
\alias{sqlToR}
\title{Convert data types from SQL to R}
\usage{
  sqlToR(data)
}
\arguments{
  \item{data}{The vector of SQL data types to convert}
}
\value{
  A vector of R data types.
}
\description{
  Takes a vector of SQL data types and and returns the
  corresponding data types in R.
}

